mitoreach_deploy 2/requirements.txt
streamlit==1.54.0
biopython==1.86
numpy==2.4.2
matplotlib==3.10.8
pandas==2.3.3
Pillow==12.1.1
